+++ src/CHANGES.md
@@ -2,6 +2,9 @@
 
 ## 0.8.135 - 2024-08-11
 * translated `/system` #1717
+* added default pin seetings for opendtufusion board
+* fixed ethernet static IP
+* fixed ethernet MAC address read back
 
 ## 0.8.134 - 2024-08-10
 * combined Ethernet and WiFi variants - Ethernet is now always included, but needs to be enabled if needed